backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.3.1

# AI API 클라이언트
anthropic==0.29.0
google-generativeai==0.7.2
openai==1.35.3
ollama==0.2.1

# 벡터 DB / 메모리
chromadb==0.5.3
sentence-transformers==3.0.1

# 데이터베이스
sqlalchemy==2.0.31
aiosqlite==0.20.0

# 이미지 처리
Pillow==10.3.0
httpx==0.27.0

# 유틸리티
python-dotenv==1.0.1
aiofiles==23.2.1
tenacity==8.3.0
rich==13.7.1
matplotlib==3.9.0
networkx==3.3